+++ mobile/src/main/java/com/github/shadowsocks/MainActivity.kt
@@ -25,11 +25,6 @@ import android.app.backup.BackupManager
 import android.content.ActivityNotFoundException
 import android.content.Intent
 import android.content.pm.ShortcutManager
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.drawable.AdaptiveIconDrawable
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.LayerDrawable
 import android.net.VpnService
 import android.os.Build
 import android.os.Bundle
@@ -38,7 +33,6 @@ import android.os.RemoteException
 import android.util.Log
 import android.view.*
 import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.content.res.AppCompatResources
 import androidx.browser.customtabs.CustomTabColorSchemeParams
 import androidx.browser.customtabs.CustomTabsIntent
 import androidx.coordinatorlayout.widget.CoordinatorLayout
@@ -60,6 +54,7 @@ import com.github.shadowsocks.preference.OnPreferenceDataStoreChangeListener
 import com.github.shadowsocks.subscription.SubscriptionFragment
 import com.github.shadowsocks.utils.Key
 import com.github.shadowsocks.utils.SingleInstanceActivity
+import com.github.shadowsocks.utils.getBitmap
 import com.github.shadowsocks.widget.ListHolderListener
 import com.github.shadowsocks.widget.ServiceButton
 import com.github.shadowsocks.widget.StatsBar
@@ -139,34 +134,21 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
     }
 
     private fun updateShortcuts() {
-        fun getBitmap(id: Int): Bitmap {
-            var drawable = AppCompatResources.getDrawable(this, id)!!
-            if (drawable is BitmapDrawable)
-                return drawable.bitmap
-            if (Build.VERSION.SDK_INT >= 26 && drawable is AdaptiveIconDrawable) {
-                drawable = LayerDrawable(arrayOf(drawable.background, drawable.foreground))
-            }
-            val bitmap = Bitmap.createBitmap(
-                    drawable.intrinsicWidth, drawable.intrinsicHeight,
-                    Bitmap.Config.ARGB_8888
-            )
-            val canvas = Canvas(bitmap)
-            drawable.setBounds(0, 0, canvas.width, canvas.height)
-            drawable.draw(canvas)
-            return bitmap
-        }
+        fun getIcon(id: Int): IconCompat = if (Build.VERSION.SDK_INT >= 26)
+            IconCompat.createWithAdaptiveBitmap(getBitmap(id))
+        else IconCompat.createWithBitmap(getBitmap(id))
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
             val shortcutManager = getSystemService<ShortcutManager>(ShortcutManager::class.java)
             val toggle = ShortcutInfoCompat.Builder(this, Shortcut.SHORTCUT_TOGGLE)
                     .setIntent(Intent(this, Shortcut::class.java).setAction(Shortcut.SHORTCUT_TOGGLE))
-                    .setIcon(IconCompat.createWithBitmap(getBitmap(R.drawable.ic_qu_shadowsocks_launcher)))
+                    .setIcon(getIcon(R.drawable.ic_qu_shadowsocks_launcher))
                     .setShortLabel(Core.app.getString(R.string.quick_toggle))
                     .build()
                     .toShortcutInfo()
             val scan = ShortcutInfoCompat.Builder(this, Shortcut.SHORTCUT_SCAN)
                     .setIntent(Intent(this, Shortcut::class.java).setAction(Shortcut.SHORTCUT_SCAN))
-                    .setIcon(IconCompat.createWithBitmap(getBitmap(R.drawable.ic_qu_camera_launcher)))
+                    .setIcon(getIcon(R.drawable.ic_qu_camera_launcher))
                     .setShortLabel(Core.app.getString(R.string.add_profile_methods_scan_qr_code))
                     .build()
                     .toShortcutInfo()
